Move navigateToPath to useMetabotChat
