+++ enterprise/frontend/src/metabase-enterprise/metabot/hooks/use-metabot-agent.ts
@@ -14,11 +14,9 @@ import {
   getMetabotId,
   getMetabotRequestId,
   getMetabotVisible,
-  getNavigateToPath,
   getToolCalls,
   resetConversation as resetConversationAction,
   retryPrompt,
-  setNavigateToPath as setNavigateToPathAction,
   setVisible as setVisibleAction,
   submitInput as submitInputAction,
 } from "../state";
@@ -49,10 +47,6 @@ export const useMetabotAgent = () => {
   const metabotRequestId = useSelector(
     getMetabotRequestId as any,
   ) as ReturnType<typeof getMetabotRequestId>;
-
-  const navigateToPath = useSelector(getNavigateToPath as any) as ReturnType<
-    typeof getNavigateToPath
-  >;
 
   const setVisible = useCallback(
     (isVisible: boolean) => dispatch(setVisibleAction(isVisible)),
@@ -134,13 +128,6 @@ export const useMetabotAgent = () => {
     [resetConversation, setVisible, promptInputRef, submitInput],
   );
 
-  const setNavigateToPath = useCallback(
-    async (navigateToPath: string) => {
-      dispatch(setNavigateToPathAction(navigateToPath));
-    },
-    [dispatch],
-  );
-
   return {
     prompt,
     setPrompt,
@@ -149,7 +136,6 @@ export const useMetabotAgent = () => {
     visible,
     messages,
     errorMessages,
-    navigateToPath,
     lastAgentMessages: useSelector(
       getLastAgentMessagesByType as any,
     ) as ReturnType<typeof getLastAgentMessagesByType>,
@@ -157,7 +143,6 @@ export const useMetabotAgent = () => {
       getIsLongMetabotConversation as any,
     ) as ReturnType<typeof getIsLongMetabotConversation>,
     resetConversation,
-    setNavigateToPath,
     setVisible,
     startNewConversation,
     submitInput,
